CarteBancaireMicroservice/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.1
py-eureka-client==0.11.7

